+++ devops/clean.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
 # Stop the infrastructure
-echo "Tests finished, stopping containers."
+echo "Stopping containers."
 docker stop mongodb
 docker stop redis