+++ linz.js
@@ -139,7 +139,7 @@ linz.hbs = require('./lib/hbs-helpers');
 linz.formtools = require('./lib/formtools');
 linz.versions = require('./lib/versions');
 linz.concurrencyControl = require('./lib/concurrency-control');
-linz.middleware = require('./middleware-public');
+linz.middleware = require('./lib/api/middleware');
 linz.api = require('./lib/api');
 linz.utils = require('./lib/utils');