+++ test/lib-defaults.test.js
@@ -28,7 +28,7 @@ test('sets a default', () => {
     expect(typeof linz.get('permissions')).toBe('function');
     expect(linz.get('login middleware')).toHaveProperty('get');
     expect(linz.get('login middleware')).toHaveProperty('post');
-    expect(linz.get('logout middleware')).toEqual({ get: [require('../middleware-public/logout')] });
+    expect(linz.get('logout middleware')).toEqual({ get: [require('../lib/api/middleware/logout')] });
     expect(linz.get('disable navigation cache')).toEqual(false);
     expect(typeof linz.get('navigationTransform')).toBe('function');
     expect(typeof linz.get('customAttributes')).toBe('function');