+++ middleware/modelVersionsCompare.js
@@ -2,6 +2,8 @@ var async = require('async'),
     deep = require('deep-diff'),
     textDiffUtils = require('diff');
 
+const { namespaceForm } = require('../lib/api/formtools');
+
 function getData (req, cb) {
 
     var versionsHelpers = require('./_modelVersionsCompare')(req);
@@ -17,85 +19,92 @@ module.exports = {
 
     get: function (req, res, next) {
 
-        var versionsHelpers = require('./_modelVersionsCompare')(req);
+        // Make sure req.linz.model.linz.formtools.form always exists.
+        namespaceForm(req)
+            .then(() => {
 
-        async.waterfall([
+                var versionsHelpers = require('./_modelVersionsCompare')(req);
 
-            (cb) => getData(req, cb),
+                async.waterfall([
 
-            (result, cb) => {
+                    (cb) => getData(req, cb),
 
-                var diffs = deep.diff(result.previous, result.latest);
+                    (result, cb) => {
 
-                if (!diffs) {
-                    // there are no change between the 2 versions, exit
-                    return cb(null, undefined);
-                }
+                        var diffs = deep.diff(result.previous, result.latest);
 
-                // get a list of changed fields and remove duplicates
-                var changedFields = diffs.map(diff => diff.path[0])
-                .filter((fieldName, index, array) =>  array.indexOf(fieldName) == index);
+                        if (!diffs) {
+                            // there are no change between the 2 versions, exit
+                            return cb(null, undefined);
+                        }
 
-                // get data ready for comparison
-                async.series({
-                    latest: done => versionsHelpers.getVersionById(req.params.revisionBId, changedFields.join(' '), true, done),
-                    previous: done => versionsHelpers.getVersionById(req.params.revisionAId, changedFields.join(' '), true, done)
-                }, cb);
+                        // get a list of changed fields and remove duplicates
+                        var changedFields = diffs.map(diff => diff.path[0])
+                        .filter((fieldName, index, array) =>  array.indexOf(fieldName) == index);
 
-            },
+                        // get data ready for comparison
+                        async.series({
+                            latest: done => versionsHelpers.getVersionById(req.params.revisionBId, changedFields.join(' '), true, done),
+                            previous: done => versionsHelpers.getVersionById(req.params.revisionAId, changedFields.join(' '), true, done)
+                        }, cb);
 
-            (result, cb) => {
+                    },
 
-                if (!result) {
-                    return cb();
-                }
+                    (result, cb) => {
 
-                var form = req.linz.model.linz.formtools.form,
-                    fieldsType = {};
+                        if (!result) {
+                            return cb();
+                        }
 
-                // populate fieldsType object
-                Object.keys(form).forEach(fieldName => fieldsType[form[fieldName].label] = form[fieldName].type);
+                        var form = req.linz.model.linz.formtools.form,
+                            fieldsType = {};
 
-                req.linz.diffs = deep.diff(result.previous, result.latest);
+                        // populate fieldsType object
+                        Object.keys(form).forEach(fieldName => fieldsType[form[fieldName].label] = form[fieldName].type);
 
-                if (!req.linz.diffs) {
-                    // there are no change between the 2 versions, exit
-                    return cb();
-                }
+                        req.linz.diffs = deep.diff(result.previous, result.latest);
 
-                // generate text diff for properties that are text fields
-                req.linz.diffs.forEach(diff => {
-                    if (diff.kind === 'E' && typeof diff.lhs === 'string' && diff.lhs.length && typeof diff.rhs === 'string' && fieldsType[diff.path[0]] === 'text') {
-                        diff.textDiff = textDiffUtils.diffWords(diff.lhs, diff.rhs);
+                        if (!req.linz.diffs) {
+                            // there are no change between the 2 versions, exit
+                            return cb();
+                        }
+
+                        // generate text diff for properties that are text fields
+                        req.linz.diffs.forEach(diff => {
+                            if (diff.kind === 'E' && typeof diff.lhs === 'string' && diff.lhs.length && typeof diff.rhs === 'string' && fieldsType[diff.path[0]] === 'text') {
+                                diff.textDiff = textDiffUtils.diffWords(diff.lhs, diff.rhs);
+                            }
+                        });
+
+                        req.linz.history = result;
+
+                        async.parallel({
+                            latest: versionsHelpers.getLatestVersionAdditionalProps,
+                            previous: versionsHelpers.getPreviousVersionAdditionalProps
+                        }, (err, record) => {
+
+                            if (err) {
+                                return cb(err);
+                            }
+
+                            req.linz.history.latest.author = record.latest.author;
+                            req.linz.history.latest.date = record.latest.date;
+                            req.linz.history.latest._id = record.latest._id;
+
+                            req.linz.history.previous.author = record.previous.author;
+                            req.linz.history.previous.date = record.previous.date;
+                            req.linz.history.previous._id = record.previous._id;
+
+                            return cb();
+
+                        });
+
                     }
-                });
 
-                req.linz.history = result;
+                ], next);
 
-                async.parallel({
-                    latest: versionsHelpers.getLatestVersionAdditionalProps,
-                    previous: versionsHelpers.getPreviousVersionAdditionalProps
-                }, (err, record) => {
-
-                    if (err) {
-                        return cb(err);
-                    }
-
-                    req.linz.history.latest.author = record.latest.author;
-                    req.linz.history.latest.date = record.latest.date;
-                    req.linz.history.latest._id = record.latest._id;
-
-                    req.linz.history.previous.author = record.previous.author;
-                    req.linz.history.previous.date = record.previous.date;
-                    req.linz.history.previous._id = record.previous._id;
-
-                    return cb();
-
-                });
-
-            }
-
-        ], next);
+            })
+            .catch(next);
 
     }