+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added the ability to create a custom form based on the model schema via `linz.api.model.generateForm()` and `linz.api.model.generateFormString()`.
 - Add the ability to edit `public` and `views` files without restarting the node process.
 - Replaced `linz.publicMiddleware.namespace` with `linz.api.middleware.setLinzNamespace()`.
+- Added persistent storage to the local development environment.
+- Added new formtool apis `namespaceForm`, `namespaceList`, `namespaceOverview`, `parseForm`.
+- Added new middleware apis `setLinzNamespace`, `setModelForm`, `setRecord`.
 
 ## v1.0.0-16.1.0 (9 August 2018)