+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fixed the datetimepicker on documentarray fields.
 - Added the ability to manually set a timezone offset using `data-utc-offset` in the date widget. This *must* be in the ISO offset format and begin with a plus (+) or minus (-) symbol with the hours and minutes having a leading 0 eg `+00:00`
 - Added the ability to provide helpText to a document array field.
+- The form DSL no longer accepts a callback function as the second parameter. The function provided must now return the form object.
 
 ## v1.0.0-16.1.0 (9 August 2018)